packages/cnes_infra/pyproject.toml
[project]
name = "cnes-infra"
version = "0.1.0"
description = "Infrastructure layer — storage adapters, ingestion clients, Alembic migrations"
requires-python = ">=3.13"
dependencies = [
    "cnes-domain",
    "sqlalchemy>=2.0",
    "psycopg[binary]>=3.2",
    "psycopg2-binary>=2.9",
    "asyncpg>=0.30",
    "alembic>=1.14",
    "python-dotenv>=1.2",
    "minio>=7.2",
]

[project.optional-dependencies]
etl = [
    "fdb>=2.0",
    "firebird-driver>=2.0",
    "firebird-base>=2.0",
    "sqlalchemy-firebird>=2.1",
    "requests>=2.32",
    "basedosdados>=2.0",
    "google-cloud-bigquery>=3.40",
    "google-cloud-storage>=3.10",
    "google-cloud-bigquery-storage>=2.36",
    "google-cloud-bigquery-connection>=1.20",
    "google-auth>=2.49",
    "pandas-gbq>=0.34",
    "pandas>=2.3",
    "pyarrow>=23.0",
    "openpyxl>=3.1",
]
otel = [
    "opentelemetry-api>=1.25",
    "opentelemetry-sdk>=1.25",
    "opentelemetry-exporter-otlp-proto-grpc>=1.25",
    "opentelemetry-instrumentation-sqlalchemy>=0.46b0",
]

[tool.uv.sources]
cnes-domain = { workspace = true }

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/cnes_infra"]
